Downgrade Baidu Map SDK version to 7.6.2 and update trace animation listener parameter type

+++ baidu-map-compose/src/main/java/com/melody/map/baidu_compose/overlay/TraceOverlay.kt
@@ -80,7 +80,7 @@ fun TraceOverlay(
     ComposeNode<TraceOverlayNode, MapApplier>(
         factory = {
             val listener = object : TraceAnimationListener {
-                override fun onTraceAnimationUpdate(p0: Float) {
+                override fun onTraceAnimationUpdate(p0: Int) {
                 }
                 override fun onTraceUpdatePosition(p0: LatLng?) {
                 }
